feat: resolve pr comment
